refactor: consolidate Help functionality into TopRightControls and remove Help component
