+++ src/routes/(app)/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
+	import { page } from '$app/stores';
 	import Chat from '$lib/components/chat/Chat.svelte';
-	import Help from '$lib/components/layout/Help.svelte';
+	// Help component removed as functionality now exists in TopRightControls
 </script>
 
-<Help />
-<Chat />
+<Chat chatIdProp={$page.params.id} />
